% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-bridges.R
\docType{data}
\name{bridges}
\alias{bridges}
\title{bridges}
\format{A tibble with 530 rows and 8 variables:
\describe{
\item{st_abbv}{The state's two-letter abbreviation}
\item{rectype}{The record type. Example: 1 = Highway}
\item{rtnum}{The bridge's route number}
\item{strcture}{The bridge's structure ID}
\item{latitude}{The bridge's latitude}
\item{longitude}{The bridge's longitude}
\item{year}{The year the bridge was built}
\item{bridgecond}{The condition of the bridge (F = Fair; G = Good; P = Poor)}
}}
\usage{
bridges
}
\description{
This data comes from the National Bridge Inventory, \url{https://www.fhwa.dot.gov/bridge/nbi.cfm}
}
\examples{
\dontrun{
bridges
}
}
\keyword{datasets}
