% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albersextra-package.R
\docType{package}
\name{ggcart-package}
\alias{ggcart}
\alias{ggcart-package}
\title{ggcart: ggcart}
\description{
A package to include Puerto Rico, the Virgin Islands and Guam in the traditional Albers maps. It also allows for easier creation of maps using lines and points.
}
\author{
\strong{Maintainer}: Heather Lendway \email{hlendway@gmail.com}

Authors:
\itemize{
  \item Greg Freedman Ellis \email{greg.freedman@gmail.com}
  \item Matt Seegebarth \email{mdseege@nppd.com}
  \item Angela Li \email{angelali921@gmail.com}
  \item Charles Minshew \email{charlespminshew@gmail.com}
  \item Karsten Maurer \email{maurerkt@miamioh.edu}
}

}
\keyword{internal}
