% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_map.R
\name{transport_sf}
\alias{transport_sf}
\title{Transport (shift, rotate and scale) an sf object}
\usage{
transport_sf(sf, ref, scale = 1, shift = c(0, 0), rotate = 0)
}
\arguments{
\item{sf}{The sf object to transport}

\item{ref}{The reference sf object}

\item{scale}{A scaling factor (defaults to 1, meaning no change in scale)}

\item{shift}{Distance to shift (TKTKTK what units is it?)}

\item{rotate}{Radians to rotate the sf object}
}
\value{
An sf object
}
\description{
Takes an sf object and a reference sf object and transports
it by scaling, shifting it and rotating it.
}
\examples{
# Transports Puerto Rico off the coast of Georgia, rotate 90 degrees
# made 4 times bigger (original Puerto Rico in red, moved in blue)
library(ggplot2)
ggplot()+
geom_sf(data=albersextra:::lower48)+
  geom_sf(data=albersextra:::puerto_rico, color = "red")+
  geom_sf(data=transport_sf(sf=albersextra:::puerto_rico,
                       ref=albersextra:::puerto_rico,
                       scale=4,
                       shift = c(-130,90),
                       rotate=pi/2), color="blue")
}
